Change the fonts task to search for commands in $PATH.
Instead of choosing the command to index the font directory based
on the location of the font directory, search $PATH for various
commands that can be used to create the indices before falling
back to the values in the environment variables.  Prefer the
indexing commands from modular X.org over native X11.

This is more resilient to differences in how the X11 directories
are laid out in the filesystem.