Update "shlibs" task to honor LDCONFIG_{ADD,REMOVE}_CMD and stdin.
Make the "shlibs" task accept LDCONFIG_ADD_CMD and
LDCONFIG_REMOVE_CMD as separate command-lines to update the cache
depending on whether the task is called to "add" or "remove".  The
default action is still to invoke a bare ${LDCONFIG} command.

Also adjust the task to accept lines from standard input of the
form:

	# SHLIB: <directory> [...]

The directory paths are searched and added to the cache.  If no
paths are given, then the action is skipped altogether.  This
allows for skipping the task actions by having no SHLIB line
passed to the task.