Always run "check" actions within the "post" meta-tasks.
Depending on the values of certain environment variables, certain
"add" or "remove" actions may not actually perform operations that
change the system and still return successfully.  Always run the
"check" actions for the corresponding main action so that any
warning messages intended for the user are written to standard
output.