Support TASK_VERBOSE environment variable to select task output.
TASK_VERBOSE is a list of package tasks for which output should be
emitted (not silent) when passed to one of the package meta-tasks
-- {pre,post}{install,remove}.  If it is the special value "all",
then all tasks will emit output.  If it is the special value
"none", then no tasks will emit output.  If it contains the special
value "default", then the default verbose tasks for that meta-task
will be appended to the given list of tasks.