Keep file ownership and mode when replacing with temporary files.
The "ocaml_findlib", "refcount_file" and "shells" tasks use the
the following algorithm to update atomically the contents of a
file:

1. Create a temporary file.
2. Write the contents to the temporary file.
3. Rename the temporary file to the correct filename.

However, the temporary files were created with umask 077, which
always made the renamed files have file mode 600.

Modify the algorithm by inserting a step between 1 and 2 where
the original file is copied over the temporary file using "cp -p",
which preserves the ownership and file mode of the original file.
This causes the final updated file to have the original ownership
and file mode.